rename  releases page
